requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
uuid==1.30
requests==2.32.0
openai==1.82.0
asyncpg==0.30.0